refactor: move memory game logic into a hook
