+++ src/features/games/types/memoryGame.types.ts
@@ -1,0 +1,9 @@
+export type Card = {
+  id: number
+  image: string
+  isFlipped: boolean
+  isMatched: boolean
+  name: string
+}
+
+export type ValidNumbers = 2 | 3 | 4 | 5 | 6 | 7 | 8
